pdf and word

+++ main.py
@@ -20,11 +20,16 @@ parser.add_argument(
 )
 parser.add_argument(
     '-c', '--cookies', 
-    help='cookies of baiduwenku.'
+    help='Cookies of baiduwenku.'
 )
 parser.add_argument(
     '-C', '--cookies_filename', 
-    help='filename of the cookies.'
+    help='Filename of the cookies.'
+)
+parser.add_argument(
+    '-t', '--temp', 
+    action='store_true',
+    help='Save temporary files into folder'
 )
 parser.add_argument(
     '-o', '--output', 
@@ -38,6 +43,7 @@ parser.add_argument(
     '-F', '--filename', 
     help='URLs in a file. One URL each line.'
 )
+
 
 args = parser.parse_args()
 cookies = ''
@@ -73,23 +79,30 @@ for url in urls:
 
     print('Success. \nParse HTML...', end='')
     page_data = re.search( r'var pageData = (.*);', html)
-    title = re.search( r'<title>(.*) - 百度文库</title>', html).group(1)
-    data = json.loads(page_data.group(1))
+    try:
+        data = json.loads(page_data.group(1))
+        title = re.search( r'<title>(.*) - 百度文库</title>', html).group(1)
+    except:
+        print('Error! It is not a Baidu Wenku document.')
+        continue
  
 
     print('Success. ')
     print('title: ', title)
-    if isinstance(data['readerInfo']['htmlUrls'], list):
+
+    
+    # PPT
+    if data['readerInfo']['tplKey'] == 'new_view':
         print('Found PPT file, prepare for downloading...', end='')
         temp_dir = url.split('?')[0].split('/')[-1][:-5]
         if os.path.exists(temp_dir):
             shutil.rmtree(temp_dir)
         os.mkdir(temp_dir)
         
-        print('Success.\nstart downloading images...', end='')
+        print('Success.\nstart downloading images...')
         imgs = data['readerInfo']['htmlUrls']
-        if len(imgs) == 3:
-            print("It seems that you provided incorrect cookies, only be able to download a part (3 page) of the file.")
+        if data['readerInfo']['page'] > len(imgs):
+            print("It seems that you provided incorrect or Non-VIP cookies, only be able to download a part of the file ({} page), not the whole file ({} page).".format(len(imgs), data['readerInfo']['page']))
         for i in range(len(imgs)):
             # TODO: theading
             percentage = (i + 1) / len(imgs) * 100
@@ -99,7 +112,7 @@ for url in urls:
             with open(os.path.join(temp_dir, str(i) + '.jpg'), 'wb') as f:
                 f.write(page.read())
 
-        print('\nmerge images to a PDF...', end='')
+        print('\nMerge images to a PDF...', end='')
         imgs = [os.path.join(temp_dir, img) for img in os.listdir(temp_dir)]
         if args.output:
             output = args.output
@@ -107,11 +120,53 @@ for url in urls:
             output = title
         with open(output + '.pdf', 'wb') as f:
             f.write(img2pdf.convert(imgs))
-        shutil.rmtree(temp_dir)
+        if not args.temp:
+            shutil.rmtree(temp_dir)
         
         print('Success.')
-        print('Saved to ' + title + '.pdf')
+        print('Saved to ' + output + '.pdf')
+
+    # PDF
+    elif data['readerInfo']['tplKey'] == 'html_view':
+        print('Found PDF or WORD file, prepare for downloading...', end='')
+        jsons = {x['pageIndex']: x['pageLoadUrl'] for x in data['readerInfo']['htmlUrls']['json']}
+        pngs = {x['pageIndex']: x['pageLoadUrl'] for x in data['readerInfo']['htmlUrls']['png']}
+        
+        temp_dir = url.split('?')[0].split('/')[-1][:-5]
+        if os.path.exists(temp_dir):
+            shutil.rmtree(temp_dir)
+        os.mkdir(temp_dir)
+        
+        print('Success.')
+
+        if data['readerInfo']['page'] > len(jsons):
+            print("It seems that you provided incorrect or Non-VIP cookies, only be able to download a part of the file ({} page), not the whole file ({} page).".format(len(jsons), data['readerInfo']['page']))
+        
+        print('start downloading jsons...')
+        for i in jsons:
+            # TODO: theading
+            percentage = i / len(jsons) * 100
+            print('\r|{}| {} / {} ({:.2f}%)'.format('=' * int(percentage // 2 - 1) + '>' + '-' * int((100 - percentage) // 2), i, len(jsons), percentage), end='')
+            request = urllib.request.Request(url=jsons[i], headers=headers)
+            page = urllib.request.urlopen(request)
+            with open(os.path.join(temp_dir, str(i) + '.json'), 'wb') as f:
+                f.write(page.read())
+        print()
+        print('start downloading pngs...')
+        for i in pngs:
+            # TODO: theading
+            percentage = i / len(pngs) * 100
+            print('\r|{}| {} / {} ({:.2f}%)'.format('=' * int(percentage // 2 - 1) + '>' + '-' * int((100 - percentage) // 2), i, len(pngs), percentage), end='')
+            request = urllib.request.Request(url=pngs[i], headers=headers)
+            page = urllib.request.urlopen(request)
+            with open(os.path.join(temp_dir, str(i) + '.png'), 'wb') as f:
+                f.write(page.read())
+        print()
+        if not args.temp:
+            shutil.rmtree(temp_dir)
+            print('Do NOT support to save into one file. Use "-t" to save temporary files.')
+
 
     else:
-        print('Do NOT support this document now.')
+        print('Do NOT support this document.')
